requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gymnasium==0.29.1
h5py==3.10.0
hydra-core==1.3.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
minari==0.4.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==23.1
pandas==2.1.4
pillow==10.2.0
portion==2.4.0
protobuf==4.25.2
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.39.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
stable-baselines3==2.2.1
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.2
tqdm==4.66.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.2
